fix(@ngtools/webpack): support sourcemaps when using substitutions option
Sourcemaps were previously potentially inaccurate when the `substitutions`
option was used due to the input source code to TypeScript differing from
the on-disk source code if any substitutions applied to the file. An intermediate
sourcemap is now created when substitutions are applied and then merged with
the sourcemap generated by TypeScript.