feat(@angular-devkit/build-angular): allow default index HTML option for application builder
The `application` builder's `index` option will now default to the value of `src/index.html`.
This value is the currently generated location for all new projects and is not typically modified.
Both the long and short form of the option `index` and `index.input` support the default value
usage. This allows the removal of the index option from the configuration of a project if preferred.
The application builder also continues to support the use of a `false` value to signify that no
index HTML file should be generated in the application's output.